requirements.txt
lyzr>=0.1.0
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
requests>=2.25.0
gradio